fix(imessage): keep resource helpers internal
